Take sys/vxworks/{os,path,pipe} from sys/unix instead.
